Skip notify if message originated from the current window
For private messages, Irssi names the window using the nick of the other party.
If the window name matches the nick, then the window is currently focused and we
don't have to send notification.

This prevents the case where you're actively participating in a private
conversation but notifications are still being sent.